+++ packages/react-native/src/evm/index.ts
@@ -1,7 +1,7 @@
-export { CoinbaseWallet, coinbaseWallet } from "./wallets/wallets/coinbase-wallet";
-export { MetaMaskWallet, metamaskWallet } from "./wallets/wallets/metamask-wallet";
-export { RainbowWallet, rainbowWallet } from "./wallets/wallets/rainbow-wallet";
-export { TrustWallet, trustWallet } from "./wallets/wallets/trust-wallet";
+export { coinbaseWallet } from "./wallets/wallets/coinbase-wallet";
+export { metamaskWallet } from "./wallets/wallets/metamask-wallet";
+export { rainbowWallet } from "./wallets/wallets/rainbow-wallet";
+export { trustWallet } from "./wallets/wallets/trust-wallet";
 
 export { ConnectWallet } from "./components/ConnectWallet";
 export { Web3Button } from "./components/Web3Button";
