+++ packages/react/src/wallet/constants/wc.ts
