+++ packages/react/src/wallet/wallets/walletConnect.ts
@@ -1,6 +1,6 @@
-import { TW_WC_PROJECT_ID } from "@thirdweb-dev/react-core";
 import { WalletConnect } from "@thirdweb-dev/wallets";
 import type { Wallet, WalletOptions } from "@thirdweb-dev/react-core";
+import { TW_WC_PROJECT_ID } from "../constants/wc";
 
 export const walletConnect = (config?: { projectId?: string }) => {
   const projectId = config?.projectId || TW_WC_PROJECT_ID;
